flight for folder delete
