+++ service/src/main/java/bio/terra/workspace/common/logging/model/ActivityFlight.java
@@ -10,7 +10,7 @@ import bio.terra.workspace.service.resource.controlled.flight.clone.dataset.Clon
 import bio.terra.workspace.service.resource.controlled.flight.clone.workspace.CloneAllResourcesFlight;
 import bio.terra.workspace.service.resource.controlled.flight.clone.workspace.CloneGcpWorkspaceFlight;
 import bio.terra.workspace.service.resource.controlled.flight.create.CreateControlledResourceFlight;
-import bio.terra.workspace.service.resource.controlled.flight.delete.DeleteControlledResourcesFlight;
+import bio.terra.workspace.service.resource.controlled.flight.delete.DeleteControlledResourceFlight;
 import bio.terra.workspace.service.resource.referenced.flight.create.CreateReferenceResourceFlight;
 import bio.terra.workspace.service.resource.referenced.flight.update.UpdateReferenceResourceFlight;
 import bio.terra.workspace.service.workspace.flight.CreateGcpContextFlightV2;
@@ -49,7 +49,7 @@ public enum ActivityFlight {
       CloneControlledGcpBigQueryDatasetResourceFlight.class.getName(),
       ActivityLogChangedTarget.RESOURCE),
   CONTROLLED_RESOURCE_DELETE_FLIGHT(
-      DeleteControlledResourcesFlight.class.getName(), ActivityLogChangedTarget.RESOURCE),
+      DeleteControlledResourceFlight.class.getName(), ActivityLogChangedTarget.RESOURCE),
   AZURE_CLOUD_CONTEXT_CREATE_FLIGHT(
       CreateAzureContextFlight.class.getName(), ActivityLogChangedTarget.AZURE_CLOUD_CONTEXT),
   AZURE_CLOUD_CONTEXT_DELETE_FLIGHT(
@@ -68,7 +68,7 @@ public enum ActivityFlight {
       CloneAllResourcesFlight.class.getName(), ActivityLogChangedTarget.RESOURCE),
   GCP_WORKSPACE_CLONE_FLIGHT(
       CloneGcpWorkspaceFlight.class.getName(), ActivityLogChangedTarget.WORKSPACE),
-  FOLDER_DELETE_FLIGHT(FolderDeleteFlight.class.getName(), ActivityLogChangedTarget.RESOURCE);
+  FOLDER_DELETE_FLIGHT(FolderDeleteFlight.class.getName(), ActivityLogChangedTarget.FOLDER);
 
   private final String flightClassName;
   private final ActivityLogChangedTarget changedTarget;
