--- service/src/main/java/bio/terra/workspace/service/folder/flights/AwaitControlledResourcesDeletionFlightStep.java
@@ -1,55 +1,0 @@
-package bio.terra.workspace.service.folder.flights;
-
-import static bio.terra.workspace.common.utils.FlightUtils.FLIGHT_POLL_CYCLES;
-import static bio.terra.workspace.common.utils.FlightUtils.FLIGHT_POLL_SECONDS;
-import static bio.terra.workspace.common.utils.FlightUtils.getRequired;
-
-import bio.terra.stairway.FlightContext;
-import bio.terra.stairway.FlightState;
-import bio.terra.stairway.FlightStatus;
-import bio.terra.stairway.Step;
-import bio.terra.stairway.StepResult;
-import bio.terra.stairway.StepStatus;
-import bio.terra.stairway.exception.DatabaseOperationException;
-import bio.terra.stairway.exception.FlightWaitTimedOutException;
-import bio.terra.stairway.exception.RetryException;
-import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.ControlledResourceKeys;
-
-public class AwaitControlledResourcesDeletionFlightStep implements Step {
-
-  @Override
-  public StepResult doStep(FlightContext context) throws InterruptedException, RetryException {
-    String flightId =
-        getRequired(
-            context.getWorkingMap(),
-            ControlledResourceKeys.DELETE_RESOURCES_FLIGHT_ID,
-            String.class);
-
-    try {
-      FlightState subflightState =
-          context.getStairway().waitForFlight(flightId, FLIGHT_POLL_SECONDS, FLIGHT_POLL_CYCLES);
-      if (FlightStatus.SUCCESS != subflightState.getFlightStatus()) {
-        // no point in retrying the await step
-        return new StepResult(
-            StepStatus.STEP_RESULT_FAILURE_FATAL,
-            subflightState
-                .getException()
-                .orElseGet(
-                    () ->
-                        new RuntimeException(
-                            String.format(
-                                "Subflight had unexpected status %s. No exception for subflight found.",
-                                subflightState.getFlightStatus()))));
-      }
-    } catch (DatabaseOperationException | FlightWaitTimedOutException e) {
-      // Retry for database issues or expired wait loop
-      return new StepResult(StepStatus.STEP_RESULT_FAILURE_RETRY, e);
-    }
-    return StepResult.getStepResultSuccess();
-  }
-
-  @Override
-  public StepResult undoStep(FlightContext context) throws InterruptedException {
-    return StepResult.getStepResultSuccess();
-  }
-}
