--- service/src/main/java/bio/terra/workspace/service/folder/flights/CreateFlightIdStep.java
@@ -1,24 +1,0 @@
-package bio.terra.workspace.service.folder.flights;
-
-import bio.terra.stairway.FlightContext;
-import bio.terra.stairway.FlightMap;
-import bio.terra.stairway.Step;
-import bio.terra.stairway.StepResult;
-import bio.terra.stairway.exception.RetryException;
-import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.ControlledResourceKeys;
-
-public class CreateFlightIdStep implements Step {
-
-  @Override
-  public StepResult doStep(FlightContext context) throws InterruptedException, RetryException {
-    FlightMap workingMap = context.getWorkingMap();
-    workingMap.put(
-        ControlledResourceKeys.DELETE_RESOURCES_FLIGHT_ID, context.getStairway().createFlightId());
-    return StepResult.getStepResultSuccess();
-  }
-
-  @Override
-  public StepResult undoStep(FlightContext context) throws InterruptedException {
-    return StepResult.getStepResultSuccess();
-  }
-}
