+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/flight/delete/DeleteSamResourceStep.java
@@ -3,9 +3,11 @@ package bio.terra.workspace.service.resource.controlled.flight.delete;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.Step;
 import bio.terra.stairway.StepResult;
+import bio.terra.workspace.common.utils.FlightUtils;
 import bio.terra.workspace.db.ResourceDao;
 import bio.terra.workspace.service.iam.AuthenticatedUserRequest;
 import bio.terra.workspace.service.iam.SamService;
+import bio.terra.workspace.service.job.JobMapKeys;
 import bio.terra.workspace.service.resource.controlled.model.ControlledResource;
 import bio.terra.workspace.service.resource.model.WsmResource;
 import java.util.UUID;
@@ -21,7 +23,6 @@ public class DeleteSamResourceStep implements Step {
   private final SamService samService;
   private final UUID workspaceUuid;
   private final UUID resourceId;
-  private final AuthenticatedUserRequest userRequest;
 
   private final Logger logger = LoggerFactory.getLogger(DeleteSamResourceStep.class);
 
@@ -29,13 +30,11 @@ public class DeleteSamResourceStep implements Step {
       ResourceDao resourceDao,
       SamService samService,
       UUID workspaceUuid,
-      UUID resourceId,
-      AuthenticatedUserRequest userRequest) {
+      UUID resourceId) {
     this.resourceDao = resourceDao;
     this.samService = samService;
     this.workspaceUuid = workspaceUuid;
     this.resourceId = resourceId;
-    this.userRequest = userRequest;
   }
 
   @Override
@@ -44,6 +43,11 @@ public class DeleteSamResourceStep implements Step {
     ControlledResource resource = wsmResource.castToControlledResource();
     // deleteControlledResource already handles duplicate deletion, so we do not need to explicitly
     // handle it inside this step.
+    final AuthenticatedUserRequest userRequest =
+        FlightUtils.getRequired(
+            flightContext.getInputParameters(),
+            JobMapKeys.AUTH_USER_INFO.getKeyName(),
+            AuthenticatedUserRequest.class);
     samService.deleteControlledResource(resource, userRequest);
     return StepResult.getStepResultSuccess();
   }
