+++ service/src/test/java/bio/terra/workspace/connected/WorkspaceConnectedTestUtils.java
@@ -37,4 +37,15 @@ public class WorkspaceConnectedTestUtils {
     assertNull(jobService.retrieveJobResult(gcpContextJobId, Object.class).getException());
     return workspaceService.getWorkspace(workspaceUuid);
   }
+
+  public void deleteWorkspaceAndGcpContext(AuthenticatedUserRequest userRequest, UUID workspaceId) {
+    Workspace workspace =
+        Workspace.builder()
+            .workspaceId(workspaceId)
+            .workspaceStage(WorkspaceStage.MC_WORKSPACE)
+            .build();
+    workspaceService.deleteGcpCloudContext(workspace, userRequest);
+
+    workspaceService.deleteWorkspace(workspace, userRequest);
+  }
 }
